feat(api): LicenseDockerCfg template function support
